# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Platform API
  version: "1.0"
  description: >
    Frontend (Next.js) ve backend (Node.js/Fastify) ajanlarının uyacağı API
    kontratı. Kaynak doğruluk: bu dosya + ../shared-types.ts + ARCHITECTURE.md.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:4000/api/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Yeni kullanıcı kaydı
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Kayıt başarılı, oturum açıldı
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessEnvelope"
                  - properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "409": { $ref: "#/components/responses/Conflict" }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: E-posta/şifre ile giriş
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Giriş başarılı
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessEnvelope"
                  - properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh cookie ile yeni access token al (rotasyonlu)
      security: []
      description: Refresh token httpOnly cookie üzerinden okunur, body gerekmez.
      responses:
        "200":
          description: Yeni access token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessEnvelope"
                  - properties:
                      data: { $ref: "#/components/schemas/AuthTokens" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      security: []
      summary: Şifre sıfırlama e-postası gönder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202": { description: İstek alındı (e-posta enumeration önlemek için her zaman 202) }
  /auth/reset-password:
    post:
      tags: [Auth]
      security: []
      summary: Şifreyi token ile sıfırla
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "204": { description: Şifre güncellendi }
        "400": { $ref: "#/components/responses/ValidationError" }
  /auth/me:
    get:
      tags: [Auth]
      summary: Mevcut oturumun kullanıcı ve üyelik bilgisi
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessEnvelope"
                  - properties:
                      data: { $ref: "#/components/schemas/AuthSession" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Kimlik doğrulama yok/geçersiz
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorEnvelope" }
...
    Conflict:
      description: Çakışma (ör. e-posta zaten kayıtlı)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorEnvelope" }
    ValidationError:
      description: Girdi doğrulama hatası
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorEnvelope" }
  schemas:
    ApiSuccessEnvelope:
      type: object
      required: [data]
      properties:
        data: {}
        meta:
          type: object
          properties:
            nextCursor: { type: string, nullable: true }
    ApiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, CONFLICT, RATE_LIMITED, INTERNAL_ERROR]
            message: { type: string }
            details:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        emailVerifiedAt: { type: string, format: date-time, nullable: true }
        role: { type: string, enum: [ADMIN, EDITOR, VIEWER] }
        createdAt: { type: string, format: date-time }
...
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        accessTokenExpiresAt: { type: string, format: date-time }
    AuthSession:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        memberships:
          type: array
          items:
            type: object
            properties:
              organizationId: { type: string, format: uuid }
              role: { type: string, enum: [OWNER, ADMIN, MEMBER] }
    AuthResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        tokens: { $ref: "#/components/schemas/AuthTokens" }
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string, minLength: 1 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
...